Task#218 initial work for getting main program flow set up
