+++ sensor/src/corefunctionality.cpp
@@ -1,3 +1,5 @@
+#include "corefunctionality.h"
+
 void activateLowPowerMode(int sleepSeconds){
     //TODO fill in
     sleep(sleepSeconds);
@@ -27,4 +29,40 @@ void transmitStoredPhosphate(){
 }
 void transmitStoredSunlight(){
     //TODO fill in
+}
+
+void turnOffWiFi() {
+    Serial.print("Turned wifi off");
+    WiFi.mode(WIFI_OFF);
+}
+
+void turnOnWifi() {
+    Serial.println();
+    if(WiFi.enableSTA(true)) {
+        Serial.print("Turned wifi on");
+    } else {
+        Serial.print("Failed to turn on wifi");
+    }
+}
+
+void connectToWifi(const char* ssid, const char* password) {
+    turnOnWifi();
+    // We start by connecting to a WiFi network
+    // TODO at some point take out serial stuff
+    Serial.println();
+    Serial.println();
+    Serial.print("Connecting to ");
+    Serial.println(ssid);
+
+    WiFi.begin(ssid, password);
+
+    while (WiFi.status() != WL_CONNECTED) {
+        delay(500);
+        Serial.print(".");
+    }
+
+    Serial.println("");
+    Serial.println("WiFi connected");
+    Serial.println("IP address: ");
+    Serial.println(WiFi.localIP());
 }
