--- sensor/src/turn_off_wifi.cpp
@@ -1,6 +1,0 @@
-#include <WiFi.h>
-
-void turnOffWiFi() {
- Serial.print("Turned wifi off");
- WiFi.mode(WIFI_OFF);
-}
