--- sensor/src/turn_on_wifi.cpp
@@ -1,10 +1,0 @@
-#include <WiFi.h>
-
-void turnOnWifi() {
-    Serial.println();
-    if(WiFi.enableSTA(true)) {
-        Serial.print("Turned wifi on");
-    } else {
-        Serial.print("Failed to turn on wifi");
-    }
-}
